requirements.txt
# Core LangChain libraries for agents
langchain
langchain-core
langchain_openai
langchainhub


# For creating the agent server
langserve
fastapi
uvicorn

# System utility tool
psutil

# Configuration and API key management
python-dotenv

# Tavily web search tool
langchain-tavily

# OpenAI SDK
openai